+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/util/HDFSCopyFromLocal.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.contrib.streaming.state;
+package org.apache.flink.streaming.util;
 
 import org.apache.flink.util.ExternalProcessRunner;
 import org.apache.hadoop.conf.Configuration;
